refactor(pager): separate pager in a different layer
