+++ src/tests.rs
@@ -18,7 +18,7 @@ impl TestData {
 #[test]
 fn test_read_write() {
     let mut data_source = Cursor::new(Vec::new());
-    let mut pager = Pager::new(1024, &mut data_source);
+    let mut pager = Bookworm::new(1024, &mut data_source);
     let test_data1 = TestData::new(10, true);
     let test_data2 = TestData::new(15, false);
     let test_data3 = TestData::new(20, true);
@@ -34,7 +34,7 @@ fn test_read_write() {
 #[test]
 fn test_iter_pages() {
     let mut data_source = Cursor::new(Vec::new());
-    let mut pager = Pager::new(1024, &mut data_source);
+    let mut pager = Bookworm::new(1024, &mut data_source);
     pager.push(&TestData::new(10, true)).unwrap();
     pager.push(&TestData::new(14, false)).unwrap();
     pager.push(&TestData::new(17, true)).unwrap();
@@ -50,7 +50,7 @@ fn test_iter_pages() {
 #[test]
 fn test_push() {
     let mut data_source = Cursor::new(Vec::new());
-    let mut pager = Pager::new(1024, &mut data_source);
+    let mut pager = Bookworm::new(1024, &mut data_source);
 
     pager.push(&TestData::new(10, true)).unwrap();
     pager.push(&TestData::new(12, false)).unwrap();
@@ -62,7 +62,7 @@ fn test_push() {
     assert_eq!(iterator.next().unwrap(), TestData::new(6, true));
 
     drop(iterator);
-    let mut pager = Pager::new(1024, &mut data_source);
+    let mut pager = Bookworm::new(1024, &mut data_source);
     pager.push(&TestData::new(18, false)).unwrap();
     let mut iterator = pager.get_iterator::<TestData>();
     assert_eq!(iterator.next().unwrap(), TestData::new(10, true));
@@ -73,7 +73,7 @@ fn test_push() {
 #[test]
 fn test_remove_page() {
     let mut data_source = Cursor::new(Vec::new());
-    let mut pager = Pager::new(32, &mut data_source);
+    let mut pager = Bookworm::new(32, &mut data_source);
     let test_data = TestData::new(10, true);
     pager.push(&test_data).unwrap();
     pager.get_page::<TestData>(0).unwrap();
